services/api/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.3.4
websockets==12.0
